+++ src/Trade360SDK.Common.Entities/Entities/Enums/BetStatusId.cs
@@ -1,8 +1,7 @@
 namespace Trade360SDK.Common.Entities.Enums
 {
-    public enum MarketStatus
+    public enum BetStatusId
     {
-        NotSet = 0,
         Open = 1,
         Suspended = 2,
         Settled = 3,